🧶️ move ipc to its own top level directory
fixes #206
