+++ app/main_process/menu.ts
@@ -1,7 +1,7 @@
 import { BrowserWindow, globalShortcut, Menu, MenuItemConstructorOptions, shell } from 'electron';
 import { createWindow } from './index';
 import { MenuItemConstructorOptionsIpc } from './types';
-import { openAbout } from './ipc/ipc';
+import { openAbout } from '../ipc/ipc_main';
 
 type ShortcutMap = Record<string, string>;
 export const menuMap: Record<number, { menu: Menu; accelerators: ShortcutMap }> = {};
